+++ Cargo.toml
@@ -8,14 +8,12 @@ repository = "https://github.com/sfackler/rust-postgres-range"
 documentation = "https://sfackler.github.io/rust-postgres-range/doc/v0.9.0/postgres_range"
 
 [features]
-with-time = ["time", "postgres-shared/with-time"]
-with-chrono = ["chrono", "postgres-shared/with-chrono"]
+with-chrono-0_4 = ["chrono-04", "postgres-types/with-chrono-0_4"]
 
 [dependencies]
-time = { version = "0.1", optional = true }
-postgres-protocol = "0.3"
-chrono = { version = "0.4.0", optional = true }
-postgres-shared = "0.4.0"
+postgres-protocol = "0.5"
+postgres-types = "0.1"
+chrono-04 = { version = "0.4", package = "chrono", optional = true }
 
 [dev-dependencies]
-postgres = { version = "0.15", features = ["with-time"] }
+postgres = { version = "0.17" }
